Codebase cleanup. Removed several files and consolidated parts of the implementation
