+++ docs/architecture/index.md
@@ -1,0 +1,1 @@
+
