+++ include/cache/commons.hpp
@@ -43,16 +43,6 @@ enum class resource_t{
   OTHER_RESOURCE,             //?
 };
 
-//__attribute__((packed)) string sucks, it is non-POD
-/**
- * @brief Context for a page. Used to determine permissions and readability in cache.
- * 
- */
-struct ctx_t{
-  uint8_t src_key[256/8];       //Key as it appears on the app definition
-  uint8_t computed_key[256/8];  //Key after the parent one has been applied. Root keys is the one of the parent. Needed to avoid nasty secrets violations
-  std::string page_tag;         //Used to identify different apps all sharing the same key. While permanent data is shared, secrets are tagged to match a specific "page"
-};
 
 }
 }
