+++ include/cache/kv-storage.hpp
@@ -1,15 +1,17 @@
 #pragma once
 
 #include "commons.hpp"
-//Permanent key/value storage. Each app class has its own.
+#include "app.hpp"
+
 
 namespace vs{
 namespace cache{
-
+    
+//Permanent key/value storage. Each app class has its own.
 struct kv_storage_t{
-    buffer_t get(const ctx_t& ctx, const char* key);
-    bool set(const ctx_t& ctx, const char* key, buffer_t value, int lifetime=-1);
-    bool reset(const ctx_t& ctx, const char* key);
+    buffer_t get(const app_ctx_t& ctx, const char* key);
+    bool set(const app_ctx_t& ctx, const char* key, buffer_t value, int lifetime=-1);
+    bool reset(const app_ctx_t& ctx, const char* key);
 };
 
 }
