--- include/config.h
@@ -1,5 +1,0 @@
-#pragma once
-//#define VS_NO_QUICKJS
-//#define VS_NO_TCC
-//#define VS_NO_LUA
-//#define VS_LOW_MEMORY
