+++ include/singletons.hpp
@@ -1,7 +1,5 @@
 #pragma once
 
-#include <utils/app-env.hpp>
-#include <utils/env.hpp>
 #include <utils/paths.hpp>
 #include <utils/policies.hpp>
 #include <cache/memory-storage.hpp>
