+++ include/ui-frame.hpp
@@ -1,8 +1,6 @@
 #pragma once
 
-#include <cstddef>
-#include <cstdint>
-#include <vector>
+
 
 namespace vs{
