+++ src/ui.cpp
@@ -5,7 +5,6 @@
 #include <ui-tree.hpp>
 
 #include "ui-frame.hpp"
-#include "utils/env.hpp"
 
 namespace vs{
